Fix: allow workgroup file deletion even if there is no revisions.

+++ src/main/java/org/linagora/linshare/core/service/impl/WorkGroupNodeServiceImpl.java
@@ -491,7 +491,9 @@ public class WorkGroupNodeServiceImpl extends GenericWorkGroupNodeServiceImpl im
 					workGroupNode.getUuid(), WorkGroupNodeType.DOCUMENT_REVISION);
 			log.addRelatedResources(getListUuid(revisions));
 			log.addAuditLightEntities(transformToAuditLightEntity(revisions));
-			revisionService.deleteAll(actor, owner, workGroup, revisions);
+			if (!revisions.isEmpty()) {
+				revisionService.deleteAll(actor, owner, workGroup, revisions);
+			}
 //			Delete the document.
 			workGroupDocumentService.delete(actor, owner, workGroup, workGroupNode);
 			if (moveDocument) {
